Add ECSettings class for editorconfig settings
This used to just be a dictionary of strings, which meant parsing logic was distributed all over the codebase and applying the settings was kind of a pain. Now we have an object with reasonable types (integers, enums, booleans, etc) that encapsulates all the parsing logic.

NOTE: this also makes the binary 64-bit only, but TM2 itself is the same (and I am dropping TM1 support), so this should be fine. (64-bit is required for auto property synthesis/)