Sign bundle for TextMate 2.x & update EditorConfig-Core (#39)
The final (non-beta) release of TextMate 2.0 added runtime hardening, which broke plugins. In TextMate 2.0.3 adds an exception for plugin bundles, though runtim hardening means they still must be signed. This adds proper signing and notarization (required in MacOS 10.15+) for the bundle.

While I was at it, I also updated EditorConfig-Core to v0.12.3 (the latest) and the Xcode project for Xcode 11 compatibility.

The new project configuration here is not really ideal -- it generates lots of warnings when building and uses an old format, but I want to get something working out the door without changing it too much. This already required a fair amount of weird finagling to get code signing to work right and to get PCRE2 (required for updated EditorConfig-Core) linking correctly.

Also updates version to v0.4.0.

Fixes #38.