Apple silicon is overdue for support (#44)
Add support for Apple ARM processors by building pcre2 from source rather than using a pre-built static library stored in the repo, and making sure we create a fat library with code for both x86 and ARM64 architectures.

This adds a new build target in Xcode for pcre2 and a special build script. It also updates the minimum required version of MacOS to 11.0 (it used to be 10.12).

I’ve had a working alpha of this out for a year(!), and this finally cleans up the way that was built.

Fixes #41.